Move language overrides file
This file doesn't do much at the moment but this keeps its location
consistent with the font overrides.
