+++ lib/gir_ffi-pango.rb
@@ -5,5 +5,5 @@ require "gir_ffi"
 GirFFI.setup :Pango
 
 require "gir_ffi-pango/font_overrides"
-require "gir_ffi-pango/language"
+require "gir_ffi-pango/language_overrides"
 require "gir_ffi-pango/version"
